reformatted to use Boostrap more correctly
Cut out a lot of fat in the CSS and used more proper rows and containers.  2 options for the special pizza menu.  plus a placeholder image. Think about reducing the table objects in to one controller with conditionals now that formatting is closer to completion. may require another data column. also consider having the algo for the table object not create the row if its empty. Still need to add nav bar