nav bar and menu lineups
Got navbar to a place im happy with. The logo is the correct size and padding.  The nav links function as intended. The menu lineup is fixed as well. The size of columns changed to make centering look better. added a bold class in CSS for the names of men items.